[Podspec + Sources + Project Settings] Fix redefinition of module `RSKImageCropper`.
Xcode 8.3 - Apple LLVM Compiler - Known Issues

When building a framework that has module maps in both the source and the install directories that define the same module, the compiler will show a redefinition message. (28638816)
Workaround: Rename the module map file in the source directory to a non-default name (the default name is module.modulemap or module.map), and set the Module Map File build setting to the renamed module map.

https://developer.apple.com/library/content/releasenotes/DeveloperTools/RN-Xcode/Chapters/Introduction.html

Close #139, fixes #141.